+++ src/txtokenlist.h
@@ -130,9 +130,9 @@ public:
 class Mactab {
 private:
     Macro **table{nullptr};    // this contains the table
-    int *   rc_table{nullptr}; // this contains the reference counts
+    long *  rc_table{nullptr}; // this contains the reference counts
     size_t  cur_rc_mac_len{0}; // size of the table.
-    int     ptr{-1};           // pointer to the first free position
+    long    ptr{-1};           // pointer to the first free position
 private:
     void rc_mac_realloc();
